pyproject.toml
[project]
name = "notion-watcher"
version = "0.1.0"
description = "A lightweight background service that monitors Notion databases for trigger events and executes automated tasks. Designed to run continuously on a Raspberry Pi server with CI/CD deployment support."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]
